requirements.txt
streamlit
requests
pandas
tqdm
yfinance
lxml
beautifulsoup4
groq
python-dateutil
dotenv